Refactored CAD code out of assembly_renderer module
